@@ -6,8 +6,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
  * Properties specific to Api Authorisation.
  * <p>
  * Properties are configured in the {@code application.yml} file.
- * See {@link io.github.jhipster.config.JHipsterProperties} for a good example.
+ * See {@link tech.jhipster.config.JHipsterProperties} for a good example.
  */
 @ConfigurationProperties(prefix = "application", ignoreUnknownFields = false)
-public class ApplicationProperties {
-}
+public class ApplicationProperties {}
