@@ -3,8 +3,6 @@ package com.boa.api.config;
 import com.fasterxml.jackson.datatype.hibernate5.Hibernate5Module;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.fasterxml.jackson.module.afterburner.AfterburnerModule;
-
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.zalando.problem.ProblemModule;
@@ -36,18 +34,10 @@ public class JacksonConfiguration {
     }
 
     /*
-     * Jackson Afterburner module to speed up serialization/deserialization.
-     */
-    @Bean
-    public AfterburnerModule afterburnerModule() {
-        return new AfterburnerModule();
-    }
-
-    /*
      * Module for serialization/deserialization of RFC7807 Problem.
      */
     @Bean
-    ProblemModule problemModule() {
+    public ProblemModule problemModule() {
         return new ProblemModule();
     }
 
@@ -55,7 +45,7 @@ public class JacksonConfiguration {
      * Module for serialization/deserialization of ConstraintViolationProblem.
      */
     @Bean
-    ConstraintViolationProblemModule constraintViolationProblemModule() {
+    public ConstraintViolationProblemModule constraintViolationProblemModule() {
         return new ConstraintViolationProblemModule();
     }
 }
