@@ -1,11 +1,9 @@
 package com.boa.api.repository;
 
 import com.boa.api.domain.Authority;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
  * Spring Data JPA repository for the {@link Authority} entity.
  */
-public interface AuthorityRepository extends JpaRepository<Authority, String> {
-}
+public interface AuthorityRepository extends JpaRepository<Authority, String> {}
