@@ -11,6 +11,5 @@ public final class AuthoritiesConstants {
 
     public static final String ANONYMOUS = "ROLE_ANONYMOUS";
 
-    private AuthoritiesConstants() {
-    }
+    private AuthoritiesConstants() {}
 }
