@@ -1,15 +1,16 @@
 package com.boa.api.config;
 
-import com.boa.api.service.MailService;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 
+import com.boa.api.service.MailService;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
 @Configuration
 public class NoOpMailConfiguration {
+
     private final MailService mockMailService;
 
     public NoOpMailConfiguration() {
