@@ -7,6 +7,4 @@ import org.springframework.stereotype.Repository;
  * Spring Data JPA repository for the {@link DateTimeWrapper} entity.
  */
 @Repository
-public interface DateTimeWrapperRepository extends JpaRepository<DateTimeWrapper, Long> {
-
-}
+public interface DateTimeWrapperRepository extends JpaRepository<DateTimeWrapper, Long> {}
